# node-service/open-api/swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Store API",
    "version": "0.1.0"
  },
  "definitions": {},
  "paths": {
    "/": {
      "get": {
        "description": "Store index page",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/order": {
      "get": {
        "description": "Get order by ID",
        "parameters": [
          {
            "name": "id",
            "type": "string",
            "required": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "post": {
        "description": "Create order",
        "parameters": [
          {
            "name": "order",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "item": {
                  "type": "string"
                }
              }
            },
            "required": true,
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/inventory": {
      "get": {
        "description": "Get orders",
        "parameters": [
          {
            "name": "id",
            "type": "string",
            "required": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "description": "Get all orders with pagination and sorting",
        "parameters": [
          {
            "name": "page",
            "type": "integer",
            "required": false,
            "in": "query",
            "default": 1
          },
          {
            "name": "per_page",
            "type": "integer",
            "required": false,
            "in": "query",
            "default": 10
          },
          {
            "name": "sort_by",
            "type": "string",
            "required": false,
            "in": "query",
            "default": "id"
          },
          {
            "name": "sort_order",
            "type": "string",
            "required": false,
            "in": "query",
            "default": "asc"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "500": {
            "description": "Error fetching orders"
          }
        }
      }
    },
    "/orders/filter": {
      "get": {
        "description": "Filter orders by date range",
        "parameters": [
          {
            "name": "start_date",
            "type": "string",
            "required": true,
            "in": "query"
          },
          {
            "name": "end_date",
            "type": "string",
            "required": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "400": {
            "description": "Invalid date range"
          },
          "500": {
            "description": "Error filtering orders"
          }
        }
      }
    },
    "/orders/search": {
      "get": {
        "description": "Search orders by order ID",
        "parameters": [
          {
            "name": "order_id",
            "type": "string",
            "required": true,
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "400": {
            "description": "Invalid order ID"
          },
          "500": {
            "description": "Error searching orders"
          }
        }
      }
    }
  }
}